pyproject.toml
[tool.poetry]
name = "testgrpc"
version = "0.1.0"
description = ""
authors = ["Beki Imanji <imanaliev@akhter.studio>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
grpcio = "^1.58.0"
grpcio-tools = "^1.58.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
